Переносит методы login,logout в SignHandler. Добавляет в лэндинг 4 последних задания. Добавляет ссылки тегам <a>.
