+++ frontend/controllers/SiteController.php
@@ -1,4 +1,5 @@
 <?php
+
 namespace frontend\controllers;
 
 use yii\web\Controller;
