+++ frontend/controllers/TasksController.php
@@ -4,19 +4,10 @@ declare(strict_types = 1);
 
 namespace frontend\controllers;
 
-use yii\web\Controller;
-
-use yii\web\NotFoundHttpException;
-use frontend\models\Tasks;
-use frontend\models\SearchTaskForm;
-use yii\helpers\ArrayHelper;
-use frontend\models\Specializations;
 use frontend\controllers\SecuredController;
-use Yii;
-
 use frontend\models\task\TaskService;
 use frontend\models\task\TaskSearchForm;
-
+use Yii;
 
 class TasksController extends SecuredController
 {
