+++ frontend/models/user/UserIdentity.php
@@ -1,0 +1,46 @@
+<?php
+
+declare(strict_types = 1);
+
+namespace frontend\models\user;
+
+use yii\db\ActiveRecord;
+use yii\web\IdentityInterface;
+
+class UserIdentity extends ActiveRecord implements IdentityInterface
+{
+    public static function tableName()
+    {
+        return 'users';
+    }
+
+    public static function findIdentity($id)
+    {
+        return self::findOne($id);
+    }
+
+    public static function findIdentityByAccessToken($token, $type = null)
+    {
+        // TODO: Implement findIdentityByAccessToken() method.
+    }
+
+    public function getId()
+    {
+        return $this->getPrimaryKey();
+    }
+
+    public function getAuthKey()
+    {
+        // TODO: Implement getAuthKey() method.
+    }
+
+    public function validateAuthKey($authKey)
+    {
+        // TODO: Implement validateAuthKey() method.
+    }
+
+    public function validatePassword($password)
+    {
+        return \Yii::$app->security->validatePassword($password, $this->password);
+    }
+}
