+++ frontend/views/layouts/main.php
@@ -1,21 +1,13 @@
 <?php
 
-/* @var $this \yii\web\View */
-/* @var $content string */
-
 use yii\helpers\Html;
 use yii\helpers\Url;
-use yii\bootstrap\Nav;
-use yii\bootstrap\NavBar;
-use yii\widgets\Breadcrumbs;
 use frontend\assets\AppAsset;
-use common\widgets\Alert;
-use yii\helpers\Url;
-
 
 AppAsset::register($this);
 $user = $this->context->user;
 ?>
+
 <?php $this->beginPage() ?>
 <!DOCTYPE html>
 <html lang="<?= Yii::$app->language ?>">
@@ -24,13 +16,8 @@ $user = $this->context->user;
     <meta charset="<?= Yii::$app->charset ?>">
     <meta http-equiv="X-UA-Compatible" content="IE=edge">
     <meta name="viewport" content="width=device-width, initial-scale=1">
-    
     <?php $this->registerCsrfMetaTags() ?>
     <title><?= Html::encode($this->title) ?></title>
-
-    <!-- <link rel="stylesheet" href="css/normalize.css">
-    <link rel="stylesheet" href="css/style.css"> -->
-
 </head>
 <body>
 <?php $this->beginBody() ?>
@@ -39,7 +26,7 @@ $user = $this->context->user;
   <header class="page-header">
     <div class="main-container page-header__container">
       <div class="page-header__logo">
-        <a href="index.html">
+        <a href="<?= Url::toRoute('tasks/index') ?>">
           <svg class="page-header__logo-image" id="Layer_2" xmlns="http://www.w3.org/2000/svg"
                viewBox="0 0 1634 646.35">
             <title>taskforce_logo2-01</title>
@@ -94,16 +81,16 @@ $user = $this->context->user;
       <div class="header__nav">
         <ul class="header-nav__list site-list">
           <li class="site-list__item">
-            <a href="#">Задания</a>
+            <a href="<?= Url::toRoute('tasks/index') ?>">Задания</a>
           </li>
           <li class="site-list__item">
-            <a href="#">Исполнители</a>
+            <a href="<?= Url::toRoute('users/index') ?>">Исполнители</a>
           </li>
           <li class="site-list__item">
-            <a href="#">Создать задание</a>
+            <a href="<?= Url::toRoute('tasks/add') ?>">Создать задание</a>
           </li>
           <li class="site-list__item">
-            <a>Мой профиль</a>
+            <a href="<?= Url::to() ?>">Мой профиль</a>
           </li>
         </ul>
       </div>
@@ -134,7 +121,7 @@ $user = $this->context->user;
       </div>
       <div class="header__account">
         <a class="header__account-photo">
-          <img src="../img/user-photo.png"
+          <img src="<?= $user->avatar ?>"
                width="43" height="44"
                alt="Аватар пользователя">
         </a>
@@ -151,7 +138,7 @@ $user = $this->context->user;
             <a href="#">Настройки</a>
           </li>
           <li>
-            <a href="<?= Url::to(['user/logout']) ?>">Выход</a>
+            <a href="<?= Url::toRoute('sign/logout') ?>">Выход</a>
           </li>
         </ul>
       </div>
@@ -180,22 +167,22 @@ $user = $this->context->user;
       <div class="page-footer__links">
         <ul class="links__list">
           <li class="links__item">
-            <a href="">Задания</a>
+            <a href="<?= Url::toRoute('tasks/index') ?>">Задания</a>
           </li>
           <li class="links__item">
             <a href="">Мой профиль</a>
           </li>
           <li class="links__item">
-            <a href="">Исполнители</a>
+            <a href="<?= Url::toRoute('users/index') ?>">Исполнители</a>
           </li>
           <li class="links__item">
-            <a href="">Регистрация</a>
+            <a href="<?= Url::toRoute('sign/signup') ?>">Регистрация</a>
           </li>
           <li class="links__item">
-            <a href="">Создать задание</a>
+            <a href="<?= Url::toRoute('tasks/add') ?>">Создать задание</a>
           </li>
           <li class="links__item">
-            <a href="">Справка</a>
+            <a href="<?= Url::to() ?>">Справка</a>
           </li>
         </ul>
       </div>
@@ -226,12 +213,6 @@ $user = $this->context->user;
     </div>
   </footer>
 </div>
-<script>
-  Dropzone.autoDiscover = false;
-
-  var dropzone = new Dropzone(".dropzone", {url: window.location.href, maxFiles: 6, uploadMultiple: true,
-    acceptedFiles: 'image/*', previewTemplate: '<a href="#"><img data-dz-thumbnail alt="Фото работы"></a>'});
-</script>
 
 <?php $this->endBody() ?>
 </body>
